refactor: rename some shared components to have the same prefix on all of them
