+++ gravitee-apim-console-webui/src/organization/configuration/policies/org-settings-platform-policies.component.ts
@@ -26,7 +26,10 @@ import { Organization } from '../../../entities/organization/organization';
 import { PlatformFlowSchema } from '../../../entities/flow/platformFlowSchema';
 import { PolicyListItem } from '../../../entities/policy';
 import { PathOperator, Step } from '../../../entities/flow/flow';
-import { GioConfirmDialogComponent, GioConfirmDialogData } from '../../../shared/components/confirm-dialog/confirm-dialog.component';
+import {
+  GioConfirmDialogComponent,
+  GioConfirmDialogData,
+} from '../../../shared/components/gio-confirm-dialog/gio-confirm-dialog.component';
 import { SnackBarService } from '../../../services-ngx/snack-bar.service';
 
 interface FlowVM {
