+++ gravitee-apim-console-webui/src/organization/configuration/tenants/org-settings-tenants.component.ts
@@ -27,7 +27,10 @@ import { GioTableWrapperFilters } from '../../../shared/components/gio-table-wra
 import { Tenant } from '../../../entities/tenant/tenant';
 import { gioTableFilterCollection } from '../../../shared/components/gio-table-wrapper/gio-table-wrapper.util';
 import { SnackBarService } from '../../../services-ngx/snack-bar.service';
-import { GioConfirmDialogComponent, GioConfirmDialogData } from '../../../shared/components/confirm-dialog/confirm-dialog.component';
+import {
+  GioConfirmDialogComponent,
+  GioConfirmDialogData,
+} from '../../../shared/components/gio-confirm-dialog/gio-confirm-dialog.component';
 
 @Component({
   selector: 'org-settings-tenants',
