+++ gravitee-apim-console-webui/src/shared/components/gio-confirm-dialog/gio-confirm-dialog.component.html
