+++ gravitee-apim-console-webui/src/shared/components/gio-form-card-group/gio-form-card-group.harness.ts
