+++ gravitee-apim-console-webui/src/shared/components/gio-form-color-input/gio-form-color-input.module.ts
