+++ gravitee-apim-console-webui/src/shared/components/gio-form-slide-toogle/gio-form-slide-toggle.component.html
